requirements.txt
django
djangorestframework
django-filter
django-jasmine